Producer Framework Mod 1.4.2 - Loading ProducersConfig on mod loading to avoid light source not being created when loading the fist save. - Fix to draw transpiler. The point of identification got a similar call after 1.5, so a change was needed to patch the right place. - Checking for producing conditions when drawing animation. - Fix to the way SubtractTimeOfDay was working. It was not converting string time to minutes, and not adjusting for the day starting time. - Changing or namespace to better organize classes. - Won't check for location condition on draw cycle, since location can't change after the producing has started.
