+++ ProducerFrameworkMod/ContentPack/ProducerConfig.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.Remoting.Messaging;
+using ProducerFrameworkMod.Controllers;
+using ProducerFrameworkMod.Utils;
 using StardewValley;
 
 namespace ProducerFrameworkMod.ContentPack
