+++ ProducerFrameworkMod/Controllers/StatsController.cs
@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using ProducerFrameworkMod.ContentPack;
+﻿using ProducerFrameworkMod.ContentPack;
 using StardewValley;
 
-namespace ProducerFrameworkMod
+namespace ProducerFrameworkMod.Controllers
 {
     public class StatsController
     {
