+++ ProducerFrameworkMod/ProducerFrameworkMod.csproj
@@ -52,7 +52,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AnimationController.cs" />
+    <Compile Include="Controllers\AnimationController.cs" />
     <Compile Include="Api\IProducerFrameworkModAPI.cs" />
     <Compile Include="Api\ProducerFrameworkModApi.cs" />
     <Compile Include="ContentPack\Animation.cs" />
@@ -69,19 +69,19 @@
     <Compile Include="ContentPack\Weather.cs" />
     <Compile Include="ContentPack\WorkingTime.cs" />
     <Compile Include="DataLoader.cs" />
-    <Compile Include="GameUtils.cs" />
-    <Compile Include="LightSourceConfigController.cs" />
+    <Compile Include="Utils\GameUtils.cs" />
+    <Compile Include="Controllers\LightSourceConfigController.cs" />
     <Compile Include="ObjectOverrides.cs" />
-    <Compile Include="ObjectUtils.cs" />
-    <Compile Include="OutputConfigController.cs" />
-    <Compile Include="ProducerController.cs" />
+    <Compile Include="Utils\ObjectUtils.cs" />
+    <Compile Include="Controllers\OutputConfigController.cs" />
+    <Compile Include="Controllers\ProducerController.cs" />
     <Compile Include="ProducerFrameworkModEntry.cs" />
-    <Compile Include="ProducerRuleController.cs" />
+    <Compile Include="Controllers\ProducerRuleController.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="NameUtils.cs" />
+    <Compile Include="Utils\NameUtils.cs" />
     <Compile Include="RestrictionException.cs" />
-    <Compile Include="SoundUtil.cs" />
-    <Compile Include="StatsController.cs" />
+    <Compile Include="Utils\SoundUtil.cs" />
+    <Compile Include="Controllers\StatsController.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="ContentPackTemplate\manifest.json">
@@ -107,6 +107,7 @@
     </None>
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\Pathoschild.Stardew.ModBuildConfig.3.1.0\build\Pathoschild.Stardew.ModBuildConfig.targets" Condition="Exists('..\packages\Pathoschild.Stardew.ModBuildConfig.3.1.0\build\Pathoschild.Stardew.ModBuildConfig.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
