+++ ProducerFrameworkMod/ProducerFrameworkModEntry.cs
@@ -19,6 +19,7 @@ namespace ProducerFrameworkMod
             Helper = helper;
             
             helper.Events.GameLoop.GameLaunched += OnGameLaunched;
+            helper.Events.GameLoop.GameLaunched += DataLoader.LoadContentPacks;
             helper.Events.GameLoop.SaveLoaded += DataLoader.LoadContentPacks;
         }
 
@@ -28,7 +29,7 @@ namespace ProducerFrameworkMod
         /// <summary>Raised after the game is launched, right before the first update tick. This happens once per game session (unrelated to loading saves). All mods are loaded and initialised at this point, so this is a good time to set up mod integrations.</summary>
         /// <param name="sender">The event sender.</param>
         /// <param name="e">The event data.</param>
-        private void OnGameLaunched(object sender, GameLaunchedEventArgs e)
+        private void OnGameLaunched(object sender, EventArgs e)
         {
             new DataLoader(Helper);
