+++ ProducerFrameworkMod/Utils/NameUtils.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using StardewModdingAPI;
+﻿using StardewModdingAPI;
 using StardewValley;
+using System.Collections.Generic;
 
-namespace ProducerFrameworkMod
+namespace ProducerFrameworkMod.Utils
 {
     internal class NameUtils
     {
