+++ ProducerFrameworkMod/manifest.json
@@ -1,7 +1,7 @@
 ﻿{
   "Name": "Producer Framework Mod",
   "Author": "Digus",
-  "Version": "1.4.1",
+  "Version": "1.4.2",
   "Description": "Framework to add rules to produce objects or change the vanilla rules.",
   "UniqueID": "Digus.ProducerFrameworkMod",
   "EntryDll": "ProducerFrameworkMod.dll",
